+++ compiler/luci/service/src/Nodes/CirclePad.test.cpp
@@ -16,7 +16,7 @@
 
 #include "luci/Service/CircleNodeClone.h"
 
-#include <luci/Service/CircleShapeInference.h>
+#include "luci/Service/CircleShapeInference.h"
 
 #include <gtest/gtest.h>
